requirements.txt
httpx
pytest
black
isort
mypy
fastapi
uvicorn
python-multipart
python-dotenv
requests
Pillow
openai==1.1.0
pdf2image
gpt_index
langchain
llama_index==0.9.15
pypdf
spacy
ragas
datasets
docx2txt
cryptography
poppler-utils
PyMuPDF